[BY-254] Move: make controller folder
