+++ lib/main.dart
@@ -1,5 +1,5 @@
-import 'package:byourside/model/authenticate/auth_controller.dart';
-import 'package:byourside/model/authenticate/nickname_controller.dart';
+import 'package:byourside/screen/authenticate/controller/auth_controller.dart';
+import 'package:byourside/screen/authenticate/controller/nickname_controller.dart';
 import 'package:byourside/screen/authenticate/info/setup_user.dart';
 import 'package:byourside/screen/authenticate/social_login/social_login.dart';
 import 'package:firebase_app_check/firebase_app_check.dart';
