+++ lib/screen/authenticate/controller/nickname_controller.dart
