+++ lib/screen/authenticate/social_login/social_login.dart
@@ -1,5 +1,5 @@
 
-import 'package:byourside/model/authenticate/auth_controller.dart';
+import 'package:byourside/screen/authenticate/controller/auth_controller.dart';
 import 'package:byourside/widget/authenticate/social_login_button.dart';
 import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
